pyproject.toml
[tool.poetry]
name = "txcode-alpaca-strategies"
version = "0.1.0"
description = ""
authors = ["TexasCoding <jeffw10278@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alpaca-py = "^0.21.0"
python-dotenv = "^1.0.1"
ta = "^0.11.0"
pandas = "^2.2.2"
tqdm = "^4.66.2"
tabulate = "^0.9.0"
requests-html = "^0.10.0"
lxml-html-clean = "^0.1.1"
openai = "^1.24.1"
slack-sdk = "^3.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
